Find unstructured tables (not registry, results, or look-up) in a notebook entry.

This function returns the locations of unstructured tables in a notebook entry. It does not show the user where any registration, results, or look-up tables might be located in a notebook entry, as information in those tables must be retrieved from other API endpoints / data warehouse tables.

Usage

list_entry_tables(entry, min_rows = NULL)

Arguments

entry

Notebook entry retrieved with the `entries$get_entry_by_id` method of the Benchling Python SDK facade object created by `benchling::connect_sdk`.

min_rows

Keep tables with this number of rows or more.

Value

List where names are days and elements are indices in the `notes` attribute that correspond to tables in the entry.

Examples

if (FALSE) {
# Initialize a Benchling Python SDK facade object.
client <- benchlingr::connect_sdk(
    tenant="https://hemoshear-dev.benchling.com",
    api_key = Sys.getenv("BENCHLING_DEV_API_KEY"))
# Retrieve notebook entries with the `entries$get_entry_by_id` method.
entry <- client$entries$get_entry_by_id("etr_T3WZTyAe")
# Find the locations of the unstructured tables in the notebook entry.
benchlingr::list_entry_tables(entry)
}