Read unstructured table in a notebook entry

Read unstructured table in a notebook entry

Usage

read_entry_table(entry, day, table_position, table_name = TRUE)

Arguments

entry

Notebook entry in JSON format.

day

Integer for the day in the notebook entry. See `find_entry_tables`.

table_position

Integer for the position of the table in the notebook entry list. See `find_entry_tables`.

table_name

bool Determines how the name of the table in the notebook entry is included in the output. If `TRUE` is selected, the table name will be added as a new column to the data frame. If `FALSE`, then the table name will be ignored.

Value

Data frame representing the unstructured table in the notebook entry.

Examples

if (FALSE) {
client <- benchling_api_auth(tenant="https://hemoshear-dev.benchling.com",
                             api_key=Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_T3WZTyAe")
table_indices <- benchlingr:::find_entry_tables(entry)
print(table_indices)
a_table <- read_entry_table(entry, day=1, table_position=2)
}